fix: resolve clang-tidy warnings in as_label, sddp_cut_io, SDDPOptions
- as_label.hpp: avoid pointer arithmetic warning (use named pointers),
  use braced init list for return
- sddp_cut_io.cpp: remove redundant size_t cast on get_numcols()
- sddp_solver.hpp: suppress padding warning on SDDPOptions (logical
  field grouping preferred over packing)

+++ source/sddp_cut_io.cpp
@@ -441,8 +441,7 @@ auto load_cuts_csv(PlanningLP& planning_lp,
 
           // Validate the column index against the current LP
           if (file_col < 0
-              || static_cast<size_t>(file_col)
-                  >= static_cast<size_t>(li_ref.get_numcols()))
+              || static_cast<size_t>(file_col) >= li_ref.get_numcols())
           {
             SPDLOG_WARN(
                 "SDDP load_cuts: column index {} out of range "
